requirements.txt
Flask
Werkzeug
PyPDF2
pdf2docx
pdf2image
Pillow
